后端升级 Spring Boot 3.5.16 → 4.1.0（含前置的依赖清理）
本提交包含两个阶段（都改了同一批 pom，无法干净拆分）：

【一、依赖清理】
- 删除零引用依赖：smartdb、concurrentlinkedhashmap-lru、poi-ooxml-full(21MB)、
  poi-scratchpad(4MB)、死的 depMgmt 条目 redisson-spring-data-27 与 springdoc 空属性
- common-api 的 pom 从 227 行降到 66 行：它整段抄了 sa-base 的 40 条依赖，
  而它本就 depend on sa-base，那份拷贝对 classpath 零影响，只是让依赖关系不可读
- 修 poi 版本错配：原先显式 poi 5.5.1 而 poi-ooxml 由 fastexcel 传递进来是 5.4.1。
  只管 poi-ooxml 不够——poi 会按「同深度先声明者胜」被 fastexcel 拽回 5.4.1，
  必须把 poi 核心一起纳入 depMgmt
- 依赖归位：qlexpress4 移到 common-api、reflections 移到 sa-admin（各自唯一使用方）
- snakeyaml / commons-compress 从 dependencies 移除、版本 pin 留在 depMgmt

【二、Spring Boot 4 迁移】
- starter 改名：spring-boot-starter-aop → -aspectj
- 换 boot4 专用 starter：mybatis-plus / sa-token
- knife4j 4.6.0 把 springdoc 钉死在 2.7.0 且无 Boot 4 版本，换回官方
  springdoc-openapi-starter-webmvc-ui 3.1.0（UI 从 /doc.html 变为 /swagger-ui/index.html）
- Jackson 2 → 3（com.fasterxml → tools.jackson），40 个源文件：
  mapper 变不可变（改 builder）、ValueSerializer/ValueDeserializer 改名、
  ContextualSerializer 接口取消、ObjectCodec 移除、writeObject→writePOJO、
  JsonProcessingException 变非受检（子类不能再 throws IOException）
- Spring 7 API：MappingJackson2HttpMessageConverter → JacksonJsonHttpMessageConverter；
  HttpComponentsClientHttpRequestFactory.setConnectTimeout 已删除，
  连接超时改由连接管理器的 ConnectionConfig 承担
- 删掉 AdminApplication 上对 UserDetailsServiceAutoConfiguration 的 exclude：
  该类在 Boot 4 挪进独立模块且不在本项目 classpath 上，而 Boot 对不存在的
  exclude 会直接启动失败
- Redisson 4：RScript.ReturnType.INTEGER → LONG；排除项 redisson-spring-data-32 已不存在
- 配置：spring.jackson 的枚举/日期开关移到 spring.jackson.datatype.*（启动期硬失败）；
  删掉 spring.data.redis.password 的空值行——Redisson 4 不再把空密码当作没密码，
  照发 AUTH 导致无密码 Redis 启动失败

验证：全模块编译通过、单测 73/73 全绿、真实启动成功（10.076s）、
      登录/验证码/swagger-ui/v3-api-docs 实测均 200。

详见 docs/后端升级白皮书-SpringBoot4.md

+++ smart-admin-api-java17-springboot3/sa-marketing/src/main/java/net/lab1024/sa/draw/runtime/DrawStockService.java
@@ -101,7 +101,7 @@ public class DrawStockService {
         Long code = redissonClient.getScript(StringCodec.INSTANCE).eval(
                 RScript.Mode.READ_WRITE,
                 DEDUCT_LUA,
-                RScript.ReturnType.INTEGER,
+                RScript.ReturnType.LONG,
                 List.of(DrawCacheKey.stock(activityCode, prizeItemId),
                         DrawCacheKey.userCount(activityCode, prizeItemId, memberName)),
                 String.valueOf(userMaxCount));
@@ -112,7 +112,7 @@ public class DrawStockService {
         redissonClient.getScript(StringCodec.INSTANCE).eval(
                 RScript.Mode.READ_WRITE,
                 ROLLBACK_LUA,
-                RScript.ReturnType.INTEGER,
+                RScript.ReturnType.LONG,
                 List.of(DrawCacheKey.stock(activityCode, prizeItemId),
                         DrawCacheKey.userCount(activityCode, prizeItemId, memberName)));
     }
